+++ hive-metastore/src/test/java/org/apache/iceberg/hive/TestHiveCatalog.java
@@ -73,6 +73,7 @@ import org.apache.iceberg.exceptions.AlreadyExistsException;
 import org.apache.iceberg.exceptions.NamespaceNotEmptyException;
 import org.apache.iceberg.exceptions.NoSuchNamespaceException;
 import org.apache.iceberg.exceptions.NoSuchTableException;
+import org.apache.iceberg.exceptions.NoSuchViewException;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableMap;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableSet;
 import org.apache.iceberg.relocated.com.google.common.collect.Maps;
@@ -153,6 +154,27 @@ public class TestHiveCatalog extends CatalogTests<HiveCatalog> {
     return new Schema(
         required(1, "id", Types.IntegerType.get(), "unique ID"),
         required(2, "data", Types.StringType.get()));
+  }
+
+  @Test
+  public void testInvalidIdentifiersWithRename() {
+    TableIdentifier invalidFrom = TableIdentifier.of(Namespace.of("l1", "l2"), "view");
+    TableIdentifier validTo = TableIdentifier.of(Namespace.of("l1"), "renamedView");
+    assertThatThrownBy(() -> catalog.renameView(invalidFrom, validTo))
+        .isInstanceOf(NoSuchViewException.class)
+        .hasMessageContaining("Invalid identifier: " + invalidFrom);
+    assertThatThrownBy(() -> catalog.renameTable(invalidFrom, validTo))
+        .isInstanceOf(NoSuchTableException.class)
+        .hasMessageContaining("Invalid identifier: " + invalidFrom);
+
+    TableIdentifier validFrom = TableIdentifier.of(Namespace.of("l1"), "view");
+    TableIdentifier invalidTo = TableIdentifier.of(Namespace.of("l1", "l2"), "view");
+    assertThatThrownBy(() -> catalog.renameView(validFrom, invalidTo))
+        .isInstanceOf(IllegalArgumentException.class)
+        .hasMessageContaining("Invalid identifier: " + invalidTo);
+    assertThatThrownBy(() -> catalog.renameTable(validFrom, invalidTo))
+        .isInstanceOf(IllegalArgumentException.class)
+        .hasMessageContaining("Invalid identifier: " + invalidTo);
   }
 
   @Test
